+++ webserver/webserver/src/main/java/io/helidon/webserver/BareRequest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.helidon.webserver.spi;
+package io.helidon.webserver;
 
 import java.net.URI;
 import java.util.List;
@@ -23,9 +23,6 @@ import java.util.Map;
 import io.helidon.common.http.DataChunk;
 import io.helidon.common.http.Http;
 import io.helidon.common.reactive.Flow;
-import io.helidon.webserver.Routing;
-import io.helidon.webserver.WebServer;
-
 
 /**
  * Bare (minimal) representation of HTTP Request. Used by {@link WebServer WebServer} implementations to invoke
