+++ webserver/webserver/src/main/java/io/helidon/webserver/Response.java
@@ -40,7 +40,6 @@ import io.helidon.common.http.Http;
 import io.helidon.common.http.MediaType;
 import io.helidon.common.reactive.Flow;
 import io.helidon.common.reactive.ReactiveStreamsAdapter;
-import io.helidon.webserver.spi.BareResponse;
 
 import io.opentracing.Span;
 import io.opentracing.SpanContext;
