+++ webserver/webserver/src/main/java/io/helidon/webserver/WebServer.java
@@ -17,16 +17,14 @@
 package io.helidon.webserver;
 
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.Map;
 import java.util.Objects;
-import java.util.ServiceLoader;
 import java.util.concurrent.CompletionStage;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
 import io.helidon.common.http.ContextualRegistry;
-import io.helidon.webserver.spi.WebServerFactory;
+import io.helidon.webserver.netty.NettyWebServer;
 
 /**
  * Represents a immutably configured WEB server.
@@ -313,8 +311,7 @@ public interface WebServer {
                 throw new IllegalStateException("No server socket configuration found for named routings: " + unpairedRoutings);
             }
 
-            WebServer result = loadFactory()
-                                    .newWebServer(configuration == null
+            WebServer result = new NettyWebServer(configuration == null
                                                           ? ServerBasicConfig.DEFAULT_CONFIGURATION
                                                           : configuration,
                                                   defaultRouting, routings);
@@ -322,15 +319,6 @@ public interface WebServer {
                 ((RequestRouting) defaultRouting).fireNewWebServer(result);
             }
             return result;
-        }
-
-        private WebServerFactory loadFactory() {
-            Iterator<WebServerFactory> implementations = ServiceLoader.load(WebServerFactory.class).iterator();
-            if (implementations.hasNext()) {
-                return implementations.next();
-            }
-
-            throw new IllegalStateException("No implementation found for SPI: " + WebServerFactory.class.getName());
         }
     }
 }
