+++ webserver/webserver/src/test/java/io/helidon/webserver/RequestTest.java
@@ -19,7 +19,6 @@ package io.helidon.webserver;
 import java.net.URI;
 
 import io.helidon.common.CollectionsHelper;
-import io.helidon.webserver.spi.BareRequest;
 
 import org.junit.jupiter.api.Test;
